# backend/src/main/java/com/vsm/okno/web/ApiController.java
package com.vsm.okno.web;

import com.vsm.okno.dto.Dto;
import com.vsm.okno.service.PlanningService;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

import java.security.Principal;
import java.util.List;
import java.util.Map;
import java.util.UUID;

// Routes map 1:1 to contracts/openapi.yaml paths; add there first if the
// contract changes, this controller just wires it to PlanningService.
@RestController
@RequestMapping("/api/v1")
public class ApiController {

    private final PlanningService service;

    public ApiController(PlanningService service) {
        this.service = service;
    }

    @PostMapping("/scenarios/import")
    public ResponseEntity<Dto.ImportResponse> importScenario(@RequestBody(required = false) Dto.ImportRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.importScenario(request));
    }

    @PostMapping("/demo/source")
    public ResponseEntity<Dto.DemoSource> importDemoSource() {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.importDemoSource());
    }

    @PostMapping("/demo/scenarios/{id}/r1-arrival")
    public Dto.DemoSource changeDemoR1Arrival(@PathVariable UUID id,
                                               @RequestBody Dto.TripTimeChange request) {
        return service.changeDemoR1Arrival(id, request.arrivalMinute());
    }

    @PostMapping("/demo/case-source")
    public ResponseEntity<Dto.CaseDataset> importCaseDataset(@RequestParam String dataset) {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.importCaseDataset(dataset));
    }

    @GetMapping("/scenarios/{id}")
    public Dto.Scenario getScenario(@PathVariable UUID id) {
        return service.getScenario(id);
    }

    @GetMapping("/scenarios/{id}/trains")
    public List<Dto.Train> getTrains(@PathVariable UUID id) {
        return service.getTrains(id);
    }

    @PostMapping("/scenarios/{id}/events")
    public ResponseEntity<Map<String, UUID>> addEvent(@PathVariable UUID id, @RequestBody Dto.ScenarioEvent event) {
        UUID newScenarioId = service.newScenarioVersion(id, event);
        return ResponseEntity.status(HttpStatus.CREATED).body(Map.of("newScenarioId", newScenarioId));
    }

    @PostMapping("/planning-jobs")
    public ResponseEntity<Dto.JobStatus> createJob(@RequestBody Dto.JobRequest request) {
        return ResponseEntity.status(HttpStatus.ACCEPTED).body(service.createJob(request));
    }

    @GetMapping("/planning-jobs/{id}")
    public Dto.JobStatus getJob(@PathVariable UUID id) {
        return service.getJob(id);
    }

    @GetMapping("/plans/{id}")
    public Dto.Plan getPlan(@PathVariable UUID id) {
        return service.getPlan(id);
    }

    @GetMapping("/plans/{id}/calendar")
    public Dto.PlanCalendar getCalendar(@PathVariable UUID id) {
        return service.getCalendar(id);
    }

    @PostMapping("/plans/{id}/approve")
    public Dto.Plan approve(@PathVariable UUID id, @RequestBody Dto.ApproveRequest request, Principal principal) {
        return service.approve(id, request, principal.getName());
    }

    @GetMapping("/auth/me")
    public Dto.Me me(Authentication auth) {
        String role = auth.getAuthorities().stream().map(a -> a.getAuthority())
                .filter(a -> a.startsWith("ROLE_")).map(a -> a.substring(5)).findFirst().orElse("");
        return new Dto.Me(auth.getName(), role);
    }

    @GetMapping("/current-plan")
    public Dto.CurrentPlan currentPlan(@RequestParam(required=false) UUID scenarioId) {
        return service.currentPlan(scenarioId);
    }

    @GetMapping("/plans/active")
    public ResponseEntity<Dto.Plan> activePlan(@RequestParam(required=false) UUID scenarioId) {
        UUID id=service.currentPlan(scenarioId).planId();
        return id==null?ResponseEntity.noContent().build():ResponseEntity.ok(service.getPlan(id));
    }

    @GetMapping("/incidents")
    public List<Dto.Incident> incidents() {
        return service.incidents();
    }

    @PostMapping("/incidents")
    public ResponseEntity<Dto.Incident> reportIncident(@RequestBody Dto.IncidentRequest request, Principal principal) {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.reportIncident(request, principal.getName()));
    }

    @GetMapping(value = "/plans/{id}/export", produces = "text/csv")
    public ResponseEntity<String> export(@PathVariable UUID id) {
        return ResponseEntity.ok().contentType(MediaType.parseMediaType("text/csv")).body(service.exportCsv(id));
    }
}