# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OKNO VSM API
  version: "0.2.0"
servers:
  - url: /api/v1
# Everything except /actuator/health requires a session (cookie JSESSIONID from
# /auth/login). State-changing requests must also send the X-XSRF-TOKEN header
# with the value of the XSRF-TOKEN cookie (set on the first response, even a 401).
# Unauthenticated calls get 401 with the standard ApiError body.
paths:
...
  /auth/me:
    get:
      summary: Current user (also used to obtain the XSRF-TOKEN cookie)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: { type: string }
                  role: { type: string, enum: [PLANNER, TECHNOLOGIST, DISPATCHER] }
        '401':
          description: Not logged in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /scenarios/import:
    post:
      summary: Import scenario
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                trains:
                  type: array
                  items:
                    $ref: '#/components/schemas/Train'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarioId: { type: string, format: uuid }
                  warnings: { type: array, items: { type: string } }
                  provenance: { type: string }
        '422':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /demo/scenarios/{id}/r1-arrival:
    post:
      summary: Change R1 arrival by creating a new scenario version and immutable snapshot
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [arrivalMinute]
              properties:
                arrivalMinute: { type: integer, minimum: 50, maximum: 60, description: "Minutes after 2028-07-01T00:00:00+03:00" }
      responses:
        '200':
          description: New saved source snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DemoSource' }
  /demo/case-source:
    post:
      summary: Load additive case-derived model datasets and capture the saved source
      description: Numeric norms come from case 06; histories, dates and trips are modelled. FULL43 has an E3 assessment, but no approved 14-day plan yet.
      parameters:
        - in: query
          name: dataset
          required: true
          schema: { type: string, enum: [FULL43, E2_6, BLOCKED6] }
      responses:
        '201':
          description: Saved case model dataset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseDataset' }
        '422':
          description: Invalid dataset or missing database profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /scenarios/{id}:
    get:
      summary: Get scenario snapshot
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  createdAt: { type: string, format: date-time }
                  provenance: { type: string }
                  trainCount: { type: integer }
  /scenarios/{id}/trains:
    get:
      summary: Get fleet for scenario
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Train' }
  /scenarios/{id}/events:
    post:
      summary: Inject event, creates new scenario version
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind: { type: string, enum: [MACHINE_DOWN, UNPLANNED_INSPECTION] }
                description: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  newScenarioId: { type: string, format: uuid }
        '422':
          description: Unknown failure kind
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /planning-jobs:
    post:
      summary: Start a planning job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobRequest' }
      description: >
        Queues the job and returns immediately (status QUEUED); the solver runs on
        a worker thread. Poll GET /planning-jobs/{id} until status is terminal.
        A repeated idempotencyKey with the SAME request body returns the existing
        job; the same key with a DIFFERENT body is 409 IDEMPOTENCY_KEY_REUSED.
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobStatus' }
        '409':
          description: idempotencyKey reused with a different request body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '422':
          description: Invalid policy / timeLimitSec (1..300) / seed / frozenUntil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /planning-jobs/{id}:
    get:
      summary: Get job status
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobStatus' }
  /plans/{id}:
    get:
      summary: Get plan detail
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
  /plans/{id}/calendar:
    get:
      summary: Calendar from the immutable snapshot and actual planner result
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: Fixed source trips and calculated service intervals
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanCalendar' }
  /plans/{id}/approve:
    post:
      summary: Approve a plan version
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expectedVersion: { type: integer }
                actorId: { type: string }
                comment: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        '409':
          description: Version conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '422':
          description: >
            PLAN_NOT_APPROVABLE - the plan carries a CRITICAL validation, the solver
            found no admissible plan, or independent validation has not passed.
            Approval requires a D2 PASS report bound to the current source snapshot,
            plan version, solver status and event fingerprint. NOT_PERFORMED blocks approval.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /current-plan:
    get:
      summary: Effective approved plan; a new draft never replaces it. Any role.
      parameters:
        - in: query
          name: scenarioId
          schema: { type: string, format: uuid }
          description: Optional immutable scenario version; selection is shared by its root.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  planId: { type: string, format: uuid, nullable: true }
  /incidents:
    get:
      summary: Dispatcher's incident log, newest first. Any role.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Incident' } }
    post:
      summary: Report a trip change, urgent maintenance or equipment failure. Logged for the planner; not fed to the solver.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [train, kind, description]
              properties:
                train: { type: string }
                kind: { type: string, enum: [TRIP_CHANGE, URGENT_MAINTENANCE, EQUIPMENT_DOWN] }
                description: { type: string, maxLength: 500 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '422':
          description: Invalid train, kind or description
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /plans/{id}/export:
    get:
      summary: Export plan as CSV
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
...
  /plans/active:
    get:
      summary: Effective human-approved Plan; all registered roles
      parameters:
        - in: query
          name: scenarioId
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }

        '204': { description: No approved plan }
components:
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        train: { type: string }
        kind: { type: string, enum: [TRIP_CHANGE, URGENT_MAINTENANCE, EQUIPMENT_DOWN] }
        description: { type: string }
        reportedBy: { type: string }
        reportedAt: { type: string, format: date-time }
    CaseDataset:
      type: object
      required: [source, dataset, trainCount, tripCount, planningSupported, warnings]
      properties:
        source: { $ref: '#/components/schemas/DemoSource' }
        dataset: { type: string, enum: [FULL43, E2_6, BLOCKED6] }
        trainCount: { type: integer }
        tripCount: { type: integer }
        planningSupported: { type: boolean }
        warnings: { type: array, items: { type: string } }
...
    DemoSource:
      type: object
      properties:
        scenarioId: { type: string, format: uuid }
        snapshotId: { type: string, format: uuid }
        snapshotHash: { type: string }
        provenance: { type: string }
    Train:
      type: object
      properties:
        id: { type: string, format: uuid }
        externalId: { type: string }
        status:
          type: string
          enum: [AVAILABLE, IN_SERVICE, READY_IDLE, HOT_RESERVE, RESERVE, MAINTENANCE, CLEANING, FAULT, TRANSFER, AWAITING_RELEASE, NOT_COMMISSIONED]
        mileageKm: { type: number }
        nextObligation: { type: string }
    JobRequest:
      type: object
      required: [scenarioId, policy]
      properties:
        scenarioId: { type: string, format: uuid }
        policy: { type: string }
        seed: { type: integer }
        timeLimitSec: { type: integer }
        frozenUntil: { type: string, format: date-time }
        idempotencyKey: { type: string }
    JobStatus:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED] }
        solverStatus: { type: string, enum: [OPTIMAL, FEASIBLE, INFEASIBLE, UNKNOWN, MODEL_INVALID] }
        planId: { type: string, format: uuid, nullable: true }
        error: { type: string, nullable: true, description: "Only when status is FAILED" }
    PlanEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        trainId: { type: string, format: uuid }
        kind: { type: string }
        startAt: { type: string }
        endAt: { type: string }
        resourceIds: { type: array, items: { type: string } }
    Validation:
      type: object
      properties:
        code: { type: string }
        severity: { type: string }
        message: { type: string }
        objectId: { type: string, nullable: true }
        startAt: { type: string, format: date-time, nullable: true }
        endAt: { type: string, format: date-time, nullable: true }
    PendingMilestone:
      type: object
      properties:
        trainId: { type: string, format: uuid }
        cycleCode: { type: string }
        nominalKm: { type: integer, format: int64 }
        remainingKm: { type: integer, format: int64 }
    ValidationReport:
      type: object
      description: Independent D2 report bound to the saved source, canonical event fingerprint and plan version. PASS applies only to its stated scope, never to actual train release.
      properties:
        schemaVersion: { type: string, enum: [d2-validation-1.0] }
        status: { type: string, enum: [NOT_PERFORMED, FAILED, PASS] }
        scope: { type: string }
        scenarioId: { type: string, format: uuid }
        sourceSnapshotId: { type: string, format: uuid, nullable: true }
        snapshotHash: { type: string }
        resultHash: { type: string, description: SHA-256 of d2-plan-events-v1 }
        policy: { type: string }
        solverStatus: { type: string }
        planVersion: { type: integer }
        checkedAt: { type: string, format: date-time }
        ruleVersion: { type: string, nullable: true }
        ruleConfirmation: { type: string, nullable: true }
        requiredServiceCount: { type: integer, nullable: true }
        pendingMilestones: { type: array, items: { $ref: '#/components/schemas/PendingMilestone' } }
        findings: { type: array, items: { $ref: '#/components/schemas/Validation' } }
    ResourceLoad:
      type: object
      properties:
        resourceId: { type: string }
        busyMinutes: { type: integer, format: int64 }
        horizonSharePercent: { type: number, description: Planned service minutes / whole horizon minutes * 100, not readiness or utilization of available shifts }
    PlanMetrics:
      type: object
      description: Counts from a D2-validated plan. No actual executed trips, readiness KPI or economic benefit is inferred.
      properties:
        scope: { type: string }
        scheduledTripCount: { type: integer }
        conflictingTripCount: { type: integer }
        requiredServiceCount: { type: integer }
        placedServiceCount: { type: integer }
        missingServiceCount: { type: integer }
        trainServiceMinutes: { type: integer, format: int64 }
        makespanMinutes: { type: integer, format: int64 }
        peakConcurrentService: { type: integer }
        resourceLoads: { type: array, items: { $ref: '#/components/schemas/ResourceLoad' } }
    Plan:
      type: object
      properties:
        id: { type: string, format: uuid }
        scenarioId: { type: string, format: uuid }
        version: { type: integer }
        status: { type: string, enum: [DRAFT, APPROVED] }
        approvedBy: { type: string, nullable: true, description: "Authenticated user who approved; the request's actorId is ignored" }
        events: { type: array, items: { $ref: '#/components/schemas/PlanEvent' } }
        validations: { type: array, items: { $ref: '#/components/schemas/Validation' } }
        snapshotHash: { type: string }
        validationStatus: { type: string, enum: [NOT_PERFORMED, FAILED, PASS] }
        validationReport: { $ref: '#/components/schemas/ValidationReport' }
        metrics:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PlanMetrics'
    PlanCalendar:
      type: object
      properties:
        scenarioId: { type: string, format: uuid }
        snapshotHash: { type: string }
        provenance: { type: string }
        policy: { type: string }
        solverStatus: { type: string }
        validationStatus: { type: string, enum: [NOT_PERFORMED, FAILED, PASS] }
        independentlyValidated: { type: boolean }
        horizonStart: { type: string, format: date-time }
        horizonEnd: { type: string, format: date-time }
        trains:
          type: array
          items: { $ref: '#/components/schemas/CalendarLane' }
        resources:
          type: array
          items: { $ref: '#/components/schemas/CalendarLane' }
        events:
          type: array
          items: { $ref: '#/components/schemas/CalendarEvent' }
    CalendarLane:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
    CalendarEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [TRIP, SERVICE] }
        trainId: { type: string, format: uuid }
        resourceId: { type: string, nullable: true }
        label: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        source: { type: string }
        reason: { type: string }
        cycleCode: { type: string, nullable: true }
        covers: { type: array, nullable: true, items: { type: string } }
        releaseOdometerKm: { type: integer, nullable: true }
        dueOdometerKm: { type: integer, nullable: true }
        dueAt: { type: string, format: date-time, nullable: true }
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
...